+++ Navigation/Application/AppDelegate.swift
@@ -25,7 +25,5 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     func application(_ application: UIApplication, didDiscardSceneSessions sceneSessions: Set<UISceneSession>) {
 
     }
-
-
 }
