+++ Navigation/Controllers/FeedControllers/InfoViewController.swift
@@ -42,7 +42,6 @@ class InfoViewController: UIViewController {
         }
     }
     
-    
     //MARK: - Private
     
     private func setupView() {
