+++ Navigation/Network Service/LoginFactory.swift
@@ -5,6 +5,8 @@
 //  Created by Вадим Виноградов on 27.04.2023.
 //
 
+import Foundation
+
 struct LoginFactory {
     func makeLoginInspector() -> LoginInspector {
         let checker = Checker()
