+++ Navigation/Network Service/UserService.swift
@@ -1,6 +1,5 @@
 //
 //  UserService.swift
-//  StorageService
 //
 //  Created by Вадим Виноградов on 10.04.2023.
 //
