+++ Navigation/Views/Cells/PhotosCollectionCell.swift
@@ -58,5 +58,4 @@ class PhotosCollectionCell: UICollectionViewCell {
             photoImageView.bottomAnchor.constraint(equalTo: contentView.bottomAnchor)
         ])
     }
-    
 }
